+++ src/main/java/com/quasiris/qsf/monitoring/MonitoringBuilder.java
@@ -62,9 +62,6 @@ public class MonitoringBuilder {
             aMonitoringDocument().
             withErrorLimitHours(error).
             withWarnLimitHours(warn).
-            withMonitoringId("processingTime").
-            withName("Processing Time").
-            withType("document").
             active(active).
             build();
         this.monitoringDocuments.add(monitoringDocument);